Fixed y limpieza de código
